unit testing improvements
