+++ translator/translate/otel/exporter/awsxray/translator_test.go
@@ -12,30 +12,11 @@ import (
 	"github.com/stretchr/testify/require"
 	"go.opentelemetry.io/collector/component"
 	"go.opentelemetry.io/collector/confmap"
-	v1 "k8s.io/api/core/v1"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/client-go/kubernetes/fake"
 
 	"github.com/aws/amazon-cloudwatch-agent/cfg/envconfig"
 	"github.com/aws/amazon-cloudwatch-agent/internal/util/testutil"
 	"github.com/aws/amazon-cloudwatch-agent/translator/translate/agent"
 	"github.com/aws/amazon-cloudwatch-agent/translator/translate/otel/common"
-)
-
-var (
-	// TestEKSDetector is used for unit testing EKS route
-	testEKSDetector = func() (common.Detector, error) {
-		cm := &v1.ConfigMap{
-			TypeMeta:   metav1.TypeMeta{Kind: "ConfigMap", APIVersion: "v1"},
-			ObjectMeta: metav1.ObjectMeta{Namespace: "kube-system", Name: "aws-auth"},
-			Data:       make(map[string]string),
-		}
-		return &common.EksDetector{Clientset: fake.NewSimpleClientset(cm)}, nil
-	}
-	// TestK8sDetector is used for unit testing k8s route
-	testK8sDetector = func() (common.Detector, error) {
-		return &common.EksDetector{Clientset: fake.NewSimpleClientset()}, nil
-	}
 )
 
 func TestTranslator(t *testing.T) {
@@ -104,7 +85,7 @@ func TestTranslator(t *testing.T) {
 				},
 				"middleware": "agenthealth/traces",
 			}),
-			detector: testEKSDetector,
+			detector: common.TestEKSDetector,
 		},
 		"WithAppSignalsEnabledK8s": {
 			input: map[string]any{
@@ -135,7 +116,7 @@ func TestTranslator(t *testing.T) {
 				},
 				"middleware": "agenthealth/traces",
 			}),
-			detector: testK8sDetector,
+			detector: common.TestK8sDetector,
 		},
 	}
 	factory := awsxrayexporter.NewFactory()
