+++ translator/translate/otel/translate_otel_test.go
@@ -10,26 +10,11 @@ import (
 	"github.com/stretchr/testify/require"
 	"go.opentelemetry.io/collector/component"
 	"go.opentelemetry.io/collector/confmap"
-	v1 "k8s.io/api/core/v1"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/client-go/kubernetes/fake"
 
 	"github.com/aws/amazon-cloudwatch-agent/translator"
 	_ "github.com/aws/amazon-cloudwatch-agent/translator/registerrules"
 	"github.com/aws/amazon-cloudwatch-agent/translator/translate/agent"
 	"github.com/aws/amazon-cloudwatch-agent/translator/translate/otel/common"
-)
-
-var (
-	// TestEKSDetector is used for unit testing EKS route
-	testEKSDetector = func() (common.Detector, error) {
-		cm := &v1.ConfigMap{
-			TypeMeta:   metav1.TypeMeta{Kind: "ConfigMap", APIVersion: "v1"},
-			ObjectMeta: metav1.ObjectMeta{Namespace: "kube-system", Name: "aws-auth"},
-			Data:       make(map[string]string),
-		}
-		return &common.EksDetector{Clientset: fake.NewSimpleClientset(cm)}, nil
-	}
 )
 
 func TestTranslator(t *testing.T) {
@@ -70,7 +55,7 @@ func TestTranslator(t *testing.T) {
 					},
 				},
 			},
-			detector: testEKSDetector,
+			detector: common.TestEKSDetector,
 		},
 		"WithAppSignalsTracesEnabled": {
 			input: map[string]interface{}{
@@ -80,7 +65,7 @@ func TestTranslator(t *testing.T) {
 					},
 				},
 			},
-			detector: testEKSDetector,
+			detector: common.TestEKSDetector,
 		},
 		"WithAppSignalsMetricsAndTracesEnabled": {
 			input: map[string]interface{}{
@@ -95,7 +80,7 @@ func TestTranslator(t *testing.T) {
 					},
 				},
 			},
-			detector: testEKSDetector,
+			detector: common.TestEKSDetector,
 		},
 	}
 	for name, testCase := range testCases {
